orders/requirements.txt
requests
django
djangorestframework
django-rest-passwordreset
pyyaml
django-cors-headers
reportLab